+++ core/components/romanesco/elements/plugins/07_computations/c_global/manageraugmentations.plugin.php
@@ -43,8 +43,39 @@ switch ($modx->event->name) {
         $controller->addHtml('<script src="/assets/components/romanescobackyard/js/manager.js?v=' . $versionJS . '"></script>');
 
         // Hide advanced ContentBlocks features
-        if ($modx->getOption('romanesco.cb_hide_advanced_features') == 1) {
-            $controller->addHtml('<script src="/assets/components/romanescobackyard/js/hide-advanced-features.js?v=' . $versionJS . '"></script>');
+        $settings = $modx->getOption('romanesco.cb_hide_settings');
+        $fields = $modx->getOption('romanesco.cb_hide_fields');
+        //$sudo = $modx->user->get('sudo');
+
+        if ($settings || $fields) {
+            $controller->addHtml("
+                <script>
+                // Wait for ContentBlocks to finish loading.
+                // Depends on arrive.js library (included in Romanesco theme).
+                $(document).arrive('#contentblocks-modal', function() {
+                    $(document).arrive('.contentblocks-modal-content', function() {
+                        let settings = '$settings';
+                        let fields = '$fields';
+
+                        // Hide settings
+                        if (settings) {
+                            settings = settings.split(',');
+                            for (const name of settings) {
+                                $('.contentblocks-modal-field [data-name=' + name + ']').parent().hide();
+                            }
+                        }
+
+                        // Hide CB fields
+                        if (fields) {
+                            fields = fields.split(',');
+                            for (const id of fields) {
+                                $('li.tooltip a[data-id=' + id + ']').closest('li').hide();
+                            }
+                        }
+                    });
+                });
+                </script>
+            ");
         }
 
         // Load Ybug widget for collecting manager feedback